Multiple changes to snapshot mechanism
Sadly LVM2 is unsupported by CloneZilla.

HTC has decided to create an alternative snapshot/rollback mechanism,
where the contents of the active partition is saved to a separate FS
to a file.

The rollback & snapshot cycle took 15 minutes on most VMs, and 8-15
minutes on the native desktops. (HTC in IOI 2023 used this between
contests to ensure that no data persists, and for rolling back after
testing out issues and their fixes.)